pkg/{distro,platform}: move BootMode to platform
Move the BootMode enum from pkg/distro to pkg/platform.  We want to
import BootMode to the disk package, which would create an import loop
with distro.
